refactor: extract prefill collective completion handler
