+++ frontier/scheduler/cluster_scheduler/base_cluster_scheduler.py
@@ -89,6 +89,7 @@ from frontier.scheduler.utils.collective_timing import (
     select_active_batch,
     validate_decode_layer_advance,
 )
+from frontier.scheduler.utils.prefill_collective import handle_prefill_sync_collective
 from frontier.scheduler.utils.execution_time_metrics import (
     build_single_layer_metrics_execution_time,
 )
@@ -3181,298 +3182,19 @@ class BaseClusterScheduler(ABC):
         *,
         direct_batch: Optional[Batch] = None,
     ):
-        """
-        Handle collective synchronization completion in prefill cluster.
-
-        This method handles completion of a canonical layer-local PREFILL wave.
-        The entry side is materialized and predicted by
-        ``_on_prefill_ep_wave_ready``; this method receives only the resulting
-        ``post_moe`` completion and advances to the next layer or handoff.
-
-        Args:
-            time: Synchronized time when the complete EP wave has reached this point
-            replica_id: ID of the replica
-            stage_id: Pipeline stage ID
-            batch_global_id: Global ID of the batch
-            sync_stage: only "post_moe" is valid here
-            layer_id: Current layer being processed
-            metrics_store: Metrics store for recording performance data
-        """
-        from frontier.events.batch_stage_end_event import BatchStageEndEvent
-        from frontier.events.prefill_sync_event import PrefillSyncEvent
-        from frontier.logger import get_cluster_logger
-        logger = get_cluster_logger(__name__, self._cluster_type.name)
-
-        # Check if this sync_stage has already been processed by another replica
-        # This can happen when multiple replicas reach the same sync point and each creates a PrefillSyncCollectiveEvent
-        if direct_batch is not None:
-            if sync_stage != "post_moe":
-                raise ValueError(
-                    "Direct dense PREFILL completion is valid only for post_moe transition"
-                )
-            sync_wait_room = {
-                # Dense completion is a full-stage operation.  Keep the
-                # internal handoff lane-free; using EP0 here would make the
-                # subsequent stage lookup accidentally depend on an EP child.
-                "batches": {None: direct_batch},
-                "arrival_times": {None: time},
-            }
-            participant_batches = sync_wait_room["batches"]
-        else:
-            if sync_stage not in self._prefill_sync_waiting_room[replica_id][stage_id][batch_global_id][layer_id]:
-                logger.debug(
-                    f"[PREFILL_SYNC][COLLECTIVE_SKIP] sync_stage={sync_stage} already processed for "
-                    f"replica={replica_id}, stage={stage_id}, batch_global_id={batch_global_id}, layer={layer_id}"
-                )
-                return []
-
-            # Get the synchronized batches and clean up waiting room
-            sync_wait_room = self._prefill_sync_waiting_room[replica_id][stage_id][batch_global_id][layer_id].pop(sync_stage)
-            participant_batches = sync_wait_room["batches"]
-
-        try:
-            participant_keys = list(participant_batches.keys())
-        except Exception:
-            participant_keys = []
-        logger.info(
-            f"[PREFILL_SYNC][COLLECTIVE] ENTER: t={time:.6f}s, replica={replica_id}, stage={stage_id}, "
-            f"layer={layer_id}, sync_stage={sync_stage}, batch_global_id={batch_global_id}, "
-            f"participant_keys={participant_keys}, "
-            f"participant_batches_type={type(participant_batches).__name__}"
+        """Delegate PREFILL collective completion to the utility handler."""
+        return handle_prefill_sync_collective(
+            self,
+            time,
+            replica_id,
+            stage_id,
+            batch_global_id,
+            sync_stage,
+            layer_id,
+            metrics_store,
+            direct_batch=direct_batch,
         )
 
-        if sync_stage != "post_moe":
-            raise ValueError(
-                "PREFILL collective completion accepts only post_moe for the "
-                "canonical per-layer EP protocol"
-            )
-
-        # Shared-domain PREFILL/MONOLITHIC MoE uses the canonical per-layer
-        # protocol. The EP key identifies the materialized wave, while each
-        # source batch retains the scheduler lane that owns its stage ticket.
-        # Keep those identities separate.
-        events = []
-
-        if sync_stage == "post_moe":
-            # post_moe is a synchronization boundary. Model execution for this layer has
-            # already been accounted in pre_moe; only layer transition / pipeline handoff
-            # remains after this collective.
-            sample_batch = select_active_batch(participant_batches)
-            if sample_batch is None:
-                logger.warning(
-                    f"[PREFILL_SYNC][COLLECTIVE] post_moe has no non-idle batch for "
-                    f"replica={replica_id}, stage={stage_id}, batch_global_id={batch_global_id}, layer={layer_id}"
-                )
-                return events
-
-            # Use one non-idle batch to derive shared layer timings for all DP lanes.
-            execution_time = self._predictor.predict_stage_execution_time(
-                sample_batch,
-                stage_id,
-                cluster_type=self._cluster_type,
-                num_layers=1,  # Single-layer granularity for prefill sync
-                layer_id=layer_id,
-                include_ffn=False,
-            )
-
-            # IMPORTANT: execution_time here is a single-layer prediction (num_layers=1)
-            # for component extraction, so it cannot be used as the stage layer count.
-            num_layers = self._predictor._num_layers_per_pipeline_stage
-            _, stage_layer_end = BaseClusterScheduler.get_pipeline_stage_layer_bounds(
-                stage_id,
-                num_layers,
-            )
-            next_layer_id = layer_id + 1
-            # Dispatch through the legacy private alias so lightweight scheduler
-            # fixtures that implement the pre-refactor hook remain compatible.
-            restored_full_stage_owners = self._restore_cohort_full_stage_owners(
-                source_batches=participant_batches,
-                replica_id=replica_id,
-                stage_id=stage_id,
-                layer_id=next_layer_id,
-                cohort_id=batch_global_id,
-                operation_kind=(
-                    "attention" if next_layer_id < stage_layer_end else "final"
-                ),
-            )
-
-            if layer_id < stage_layer_end - 1:
-                # Not the last layer, continue to next layer by paying next-layer attention.
-                next_layer_execution_time = self._predictor.predict_stage_execution_time(
-                    sample_batch,
-                    stage_id,
-                    cluster_type=self._cluster_type,
-                    num_layers=1,
-                    layer_id=next_layer_id,
-                    include_ffn=False,
-                )
-                attention_time_ms = next_layer_execution_time.get_single_layer_attention_scope_time()
-                attention_time = attention_delay_seconds(next_layer_execution_time)
-                total_time_to_next_sync = attention_time
-
-                for replica_local_id, batch in participant_batches.items():
-                    if batch.is_idle:
-                        logger.info(
-                            f"[PREFILL_SYNC][IDLE_SKIP] Skip next-layer pre_moe scheduling for idle batch {batch.id} "
-                            f"(replica={replica_id}, replica_local_id={replica_local_id}, "
-                            f"layer={layer_id})"
-                        )
-                        continue
-                    component_ledger = getattr(
-                        batch,
-                        "_prefill_model_execution_components_ms_by_stage",
-                        None,
-                    )
-                    if (
-                        not isinstance(component_ledger, dict)
-                        or stage_id not in component_ledger
-                        or not isinstance(component_ledger[stage_id], list)
-                    ):
-                        raise ValueError(
-                            "missing PREFILL model-execution component ledger: "
-                            f"replica={replica_id}, replica_local_id={replica_local_id}, "
-                            f"stage={stage_id}, layer={layer_id}, "
-                            f"batch_global_id={batch_global_id}, batch_id={batch.id}"
-                        )
-                    component_ledger[stage_id].append(attention_time_ms)
-                    if not restored_full_stage_owners:
-                        self.transition_stage_admission_for_layer(
-                            batch,
-                            stage_id=stage_id,
-                            layer_id=next_layer_id,
-                            operation_kind="attention",
-                            scope=FULL_STAGE_WORLD,
-                        )
-                    events.append(
-                        PrefillSyncEvent(
-                            time + total_time_to_next_sync,
-                            replica_id,
-                            stage_id,
-                            batch,
-                            getattr(
-                                batch,
-                                "_stage_owner_replica_local_id",
-                                None,
-                            ),
-                            "pre_moe",
-                            next_layer_id,
-                            total_time_to_next_sync,
-                            cluster_type=self._cluster_type,
-                        )
-                    )
-            else:
-                # Last layer completed, proceed to pipeline communication.
-                # Idle batches are synthetic synchronization placeholders and should not
-                # create stage-end / kv-transfer events in PREFILL.
-                for replica_local_id, batch in participant_batches.items():
-                    if batch.is_idle:
-                        logger.info(
-                            f"[PREFILL_SYNC][IDLE_SKIP] Skip final stage-end for idle batch {batch.id} "
-                            f"(replica={replica_id}, replica_local_id={replica_local_id}, "
-                            f"layer={layer_id})"
-                        )
-                        continue
-
-                    stage_identity = getattr(
-                        batch,
-                        "_stage_owner_replica_local_id",
-                        None,
-                    )
-                    stage_scheduler = self.get_replica_stage_scheduler(
-                        replica_id, stage_identity, stage_id
-                    )
-                    is_last_stage = stage_scheduler.is_last_stage
-                    pipeline_time = execution_time.pipeline_time * 1e-3
-                    if not hasattr(batch, "_prefill_stage_start_time"):
-                        raise ValueError(
-                            "missing PREFILL stage start time: "
-                            f"replica={replica_id}, replica_local_id={replica_local_id}, "
-                            f"stage={stage_id}, layer={layer_id}, "
-                            f"batch_global_id={batch_global_id}, batch_id={batch.id}"
-                        )
-                    original_start_time = batch._prefill_stage_start_time
-                    component_ledger = getattr(
-                        batch,
-                        "_prefill_model_execution_components_ms_by_stage",
-                        None,
-                    )
-                    if (
-                        not isinstance(component_ledger, dict)
-                        or stage_id not in component_ledger
-                        or not isinstance(component_ledger[stage_id], list)
-                        or not component_ledger[stage_id]
-                    ):
-                        raise ValueError(
-                            "missing PREFILL model-execution component ledger: "
-                            f"replica={replica_id}, replica_local_id={replica_local_id}, "
-                            f"stage={stage_id}, layer={layer_id}, "
-                            f"batch_global_id={batch_global_id}, batch_id={batch.id}"
-                        )
-                    final_timing = prepare_prefill_final_timing(
-                        execution_time,
-                        component_ledger[stage_id],
-                        time,
-                        original_start_time,
-                    )
-                    actual_model_execution_time = (
-                        final_timing.explicit_model_time + final_timing.pipeline_time
-                    )
-                    total_final_time = final_timing.total_time
-                    completion_time = final_timing.completion_time
-                    actual_execution_time = final_timing.actual_execution_time
-
-                    # Create batch stage for metrics
-                    batch_stage, _ = stage_scheduler.predict_and_create_stage(batch, skip_get_execution_time=True)
-
-                    # Schedule the batch stage with the original start time
-                    batch_stage.on_schedule(original_start_time)
-
-                    # Override with correct values:
-                    # - execution_time: actual wall-clock time including sync overhead
-                    # - model_execution_time: pure model computation time (no CPU overhead)
-                    batch_stage.override_execution_time(actual_execution_time)
-                    batch_stage.override_model_execution_time(
-                        actual_model_execution_time
-                    )
-
-                    # TODO: CHECK OVERIDE LOGIC AND METRIC LOGIC HERE
-                    # Create a corrected ExecutionTime object for metrics recording.
-                    # For mixed-layer MoE models, augment trace-only dense MLP components
-                    # from a representative dense layer so op-level traces include both
-                    # dense and MoE FFN scopes.
-                    corrected_execution_time = (
-                        self._create_prefill_corrected_execution_time_for_metrics(
-                            sample_batch,
-                            stage_id,
-                            execution_time,
-                            actual_execution_time,
-                            original_start_time,
-                        )
-                    )
-
-                    # Record metrics with correct start time and corrected execution time
-                    metrics_store.on_replica_stage_schedule(
-                        original_start_time, replica_id, stage_id, batch_stage, corrected_execution_time,
-                        self._cluster_type, stage_identity
-                    )
-
-                    # Schedule batch stage end
-                    events.append(BatchStageEndEvent(
-                        completion_time, replica_id, stage_id, is_last_stage,
-                        batch, batch_stage, self._cluster_type, stage_identity
-                    ))
-
-                    # Check if KV cache transfer should be triggered
-                    if self._should_trigger_kv_transfer(batch):
-                        kv_transfer_events = self._create_kv_transfer_events(
-                            completion_time, batch, replica_id, stage_identity
-                        )
-                        events.extend(kv_transfer_events)
-
-                    # Note: _prefill_stage_start_time cleanup moved to BatchStageEndEvent
-                    # to ensure proper detection of completed prefill sync batches
-
-        return events
 
     def _create_prefill_corrected_execution_time_for_metrics(
         self,
